requirements.txt
amqp==5.2.0
anyio==4.2.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.3.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==0.7
cron_descriptor==1.4.0
cryptography==42.0.1
cssselect2==0.7.0
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.7
django-allauth==0.54.0
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.5
django-health-check==3.18.0
django-model-utils==4.3.1
django-prometheus==2.3.1
django-redis==5.3.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.4
environ==1.0
et-xmlfile==1.1.0
filelock==3.13.1
flower==2.0.1
h11==0.14.0
hiredis==2.2.3
html5lib==1.1
httptools==0.6.1
humanize==4.9.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
lxml==5.1.0
numpy==1.26.3
oauthlib==3.2.2
openpyxl==3.1.2
orm-choices==1.0.0
oscrypto==1.3.0
pandas==2.2.0
Pillow==10.0.1
pipenv==2023.11.15
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.0
referencing==0.32.1
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.2.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.1.0
uvicorn==0.23.2
vine==5.1.0
virtualenv==20.25.0
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
whitenoise==6.5.0
xhtml2pdf==0.2.14
